feat(modularization): S-6 -- extract agents view + channel management to web/modules/agents.js
Extract ~3000 lines of agents/channel frontend from the monolithic app.js
into a dedicated ES module. Covers: agent grid/cards render, agent detail
panel, channel setup/pending/approve/allowed/invites UI, agent wizard,
start/stop/restart UI, team org-chart (tree view), and the grid/tree
view toggle with alias side-effect.

Key design decisions:
- DI pattern (initAgents) for openModal/closeModal/loadSkills/openTerminalModal/
  openConversationModal/setChatSelectedAgent -- keeps the modal lifecycle and
  chat state in app.js without circular imports.
- Avatar cache-bust epoch (_avatarEpoch, bumpAvatarEpoch) moves to agents.js;
  avatarBust() is re-exported so remaining app.js sections (skills, memories)
  share the same epoch.
- federatedPeerStatus moved to agents.js with exported get/set; Messages page
  (loadChatAgentList) and federation page update it via setFederatedPeerStatus.
- _agentsActiveView exported as get/setAgentsActiveView (alias callback) and
  setAgentsView (DOM update trigger).
- Test files updated: openAgentDetail and renderFederatedAgentCards now read
  from web/modules/agents.js.
