+++ src/__tests__/main-agent-detail-guards.test.ts
@@ -8,7 +8,8 @@ import type { RouteContext } from '../web/routes/types.js'
 
 const __dirname = dirname(fileURLToPath(import.meta.url))
 const agentsSource = readFileSync(join(__dirname, '..', 'web', 'routes', 'agents.ts'), 'utf8')
-const appSource = readFileSync(join(__dirname, '..', '..', 'web', 'app.js'), 'utf8')
+// openAgentDetail was extracted to web/modules/agents.js in S-6 modularization.
+const appSource = readFileSync(join(__dirname, '..', '..', 'web', 'modules', 'agents.js'), 'utf8')
 
 function sourceBetween(source: string, startMarker: string, endMarker: string): string {
   const start = source.indexOf(startMarker)
